Add parsing tests and refactor
