+++ package.json
@@ -8,5 +8,8 @@
     "dotenv": "^16.5.0",
     "is-fullwidth-code-point": "^5.0.0",
     "openai": "^4.103.0"
+  },
+  "scripts": {
+    "test": "node test/parseQuotes.test.js"
   }
 }
